Dedicated Unity Workspaces
I have create an empty 3D Unity sandbox for everyone to play around with and experiment in. ヾ(＠⌒ー⌒＠)ノ

In the scenes folder you will find a folder dedicated just for you. It will have your name on it along with a few folders & a personal scene. Try to keep as much of what you do in that personal scene and dedicated folder as you can. This, paired with using your own branch, will help to minimise conflicts and smooth the development process. (～o￣3￣)～

~(￣▽￣)~*  For now, feel free to create a new branch and start experimenting with some of your amazing ideas. If you need some inspiration check the GitHub issues on the left hand side of gitkraken. I've already created some tasks based on the ideas we had earlier. If you decide to do any of them, be sure to assign yourself to it on the issues page so everyone can see it's being worked on so nobody ends up wasting their time doing already done work.